chore: Signup response
signup.js, signup.php : The response will be a JSON object with the 'ok' key (it will be true or false) for telling the client that there was an error at the sign up process, and the 'error' key (that will be our error code) to define from where the error comes and change the UI accordingly.

db.sql : Changed the privileges of the PHP user access in the database, in order to prevent some malicious injection, like DROP TABLE etc.

Security has to be improved : in javascript and server-side.