chore: Login and Signup
index.html : Some bootstrap style changes.

signup.php : Adding a token that will be used if the user wants to stay connected after they exit the website. It will be stored in the database.

db.sql : Added token and token_time columns in order to keep stored the token in the database.

login.php : The script for the login, not finished. It send a json object response, so if the user don't complete correctly the fields, in the client side we can show the user what is the problem with they credentials (if it don't compromise security).

Security has to be improved : in javascript and server-side.